+++ docs/history.md
@@ -3,9 +3,10 @@
 <!-- cspell:ignore antizoubilamaka,Cran,Erbear,grihapro,Neonomi,sanstheazmo,slvn -->
 <!-- markdownlint-disable MD001 -->
 
-### _v0.10.2_ - October 23rd, 2023
+### _v0.10.0_ - October 23rd, 2023
 
-- !CHAR ORDER
+- The second random character will now always be unlocked after the three stat collectibles and before any other unlocks.
+- Each character is now guaranteed to unlock another random character from defeating It Lives.
 - Gehenna is now flipped. (However, rooms with Ball and Chains are not flipped to avoid bugs.)
 - Fixed the bug where the following rooms were flipped incorrectly:
   - Caves: #11058, #21058